feat(xmonad): restore borders for most layouts
As of xmonad/xmonad-contrib#682, the X.L.VoidBorders module has gained
the ability to restore borders for other layouts through the
normalBorders layout modifier.

This commit switches from the locally-implemented X.L.ZeroBorders to the
upstream module submitted to xmonad-contrib, and ensures the borders are
restored on layouts not using the voidBorders modifier.